refactor: share story parser
