+++ branching_novel.py
@@ -69,318 +69,11 @@ from tkinter import ttk, filedialog, messagebox
 from i18n import tr, set_language, set_language_from_file
 
 from auto_update import check_for_update
+from story_parser import Choice, Action, Branch, Chapter, Story, ParseError, StoryParser
 
 
 APP_NAME = "Branching Novel GUI"
 INSTALLER_NAME = "BranchingNovelGUI-Online-Setup.exe"
-
-
-@dataclass
-class Choice:
-    text: str
-    target_id: str
-    condition: Optional[str] = None
-
-
-@dataclass
-class Action:
-    op: str  # e.g. 'set', 'add', 'sub', 'mul', 'div', 'floordiv', 'mod', 'pow', 'expr'
-    var: str
-    value: Union[int, float, bool, str]
-
-
-@dataclass
-class Branch:
-    """A branch belonging to an existing chapter."""
-
-    branch_id: str
-    title: str
-    chapter_id: str = ""
-    paragraphs: List[str] = field(default_factory=list)
-    choices: List[Choice] = field(default_factory=list)
-    actions: List[Action] = field(default_factory=list)
-
-
-@dataclass
-class Chapter:
-    """A chapter containing multiple branches."""
-
-    chapter_id: str
-    title: str
-    branches: Dict[str, Branch] = field(default_factory=dict)
-
-
-@dataclass
-class Story:
-    title: str = "Untitled"
-    start_id: Optional[str] = None  # start branch ID
-    ending_text: str = "The End"
-    show_disabled: bool = False
-    chapters: Dict[str, Chapter] = field(default_factory=dict)
-    branches: Dict[str, Branch] = field(default_factory=dict)
-    variables: Dict[str, Union[int, float, bool, str]] = field(default_factory=dict)
-
-    def get_branch(self, bid: str) -> Optional[Branch]:
-        return self.branches.get(bid)
-
-    def get_chapter(self, cid: str) -> Optional[Chapter]:
-        return self.chapters.get(cid)
-
-
-class ParseError(Exception):
-    pass
-
-
-class StoryParser:
-    """Simple grammar parser.
-
-    - Comments are not supported; blank lines separate paragraphs.
-    - Within a chapter, paragraph and choice lines may be interleaved.
-    - Choice lines start with ``* `` and must contain ``->``.
-    """
-
-    def parse(self, text: str) -> Story:
-        text = text.lstrip("\ufeff")
-        lines = [ln.lstrip("\ufeff") for ln in text.splitlines()]
-        story = Story()
-        current_chapter: Optional[Chapter] = None
-        current_branch: Optional[Branch] = None
-        paragraph_buffer: List[str] = []
-
-        # temporarily store metadata
-        for idx, raw in enumerate(lines):
-            line = raw.strip()
-
-            # process metadata
-            if line.startswith("@title:"):
-                story.title = line[len("@title:"):].strip() or "Untitled"
-                continue
-            if line.startswith("@start:"):
-                story.start_id = line[len("@start:"):].strip() or None
-                continue
-            if line.startswith("@ending:"):
-                story.ending_text = line[len("@ending:"):].strip() or "The End"
-                continue
-            if line.startswith("@show-disabled:"):
-                val = line[len("@show-disabled:"):].strip().lower()
-                story.show_disabled = val in ("true", "1", "yes", "on")
-                continue
-
-        # actual parsing loop
-        i = 0
-        while i < len(lines):
-            raw = lines[i]
-            line = raw.rstrip("\n")
-            stripped = line.strip()
-            i += 1
-
-            if stripped.startswith("@chapter"):
-                # chapter declaration
-                current_branch = None
-                current_chapter = self._parse_chapter_decl(stripped)
-                if current_chapter.chapter_id in story.chapters:
-                    raise ParseError(f"Duplicate chapter id: {current_chapter.chapter_id}")
-                story.chapters[current_chapter.chapter_id] = current_chapter
-                continue
-
-            # other metadata handled in first pass
-            if (
-                stripped.startswith("@title:")
-                or stripped.startswith("@start:")
-                or stripped.startswith("@ending:")
-                or stripped.startswith("@show-disabled:")
-            ):
-                continue
-
-            # start of branch
-            if stripped.startswith("#"):
-                if current_chapter is None:
-                    raise ParseError("Branch defined outside of a chapter.")
-                if current_branch is not None and paragraph_buffer:
-                    merged = self._merge_paragraph_buffer(paragraph_buffer)
-                    current_branch.paragraphs.extend(merged)
-                    paragraph_buffer.clear()
-
-                current_branch = self._parse_branch_header(stripped, current_chapter.chapter_id)
-                if current_branch.branch_id in story.branches:
-                    raise ParseError(f"Duplicate branch id: {current_branch.branch_id}")
-                story.branches[current_branch.branch_id] = current_branch
-                current_chapter.branches[current_branch.branch_id] = current_branch
-                continue
-
-            # blank line indicates paragraph boundary
-            if stripped == "":
-                if current_branch is not None:
-                    merged = self._merge_paragraph_buffer(paragraph_buffer)
-                    current_branch.paragraphs.extend(merged)
-                    paragraph_buffer.clear()
-                continue
-
-            # state modification directive
-            if stripped.startswith("!"):
-                action = self._parse_action_line(stripped)
-                if current_branch is None:
-                    if action.op == "set":
-                        story.variables[action.var] = action.value
-                    elif action.op == "expr":
-                        try:
-                            story.variables[action.var] = eval(action.value, {}, dict(story.variables))
-                        except Exception:
-                            raise ParseError("Invalid expression for initial variable.")
-                    else:
-                        raise ParseError("State change found outside of any branch.")
-                else:
-                    current_branch.actions.append(action)
-                continue
-
-            # choice line
-            if stripped.startswith("* "):
-                if current_branch is None:
-                    raise ParseError("Choice found outside of any branch.")
-                choice = self._parse_choice_line(stripped)
-                current_branch.choices.append(choice)
-                continue
-
-            # normal body line
-            if current_branch is None:
-                raise ParseError("Found narrative text outside of a branch. Add a branch header starting with '#'.")
-            paragraph_buffer.append(line)
-
-        # end-of-file handling
-        if current_branch is not None and paragraph_buffer:
-            merged = self._merge_paragraph_buffer(paragraph_buffer)
-            current_branch.paragraphs.extend(merged)
-            paragraph_buffer.clear()
-
-        # if start branch not specified, use first branch
-        if story.start_id is None:
-            if story.branches:
-                story.start_id = next(iter(story.branches.keys()))
-            else:
-                raise ParseError("No branches found in story.")
-
-        # basic check for missing targets (non-existent targets also checked at runtime)
-        return story
-
-    def _merge_paragraph_buffer(self, buffer: List[str]) -> List[str]:
-        """
-        Convert consecutive body lines into a list of paragraphs separated by blank lines.
-        If a blank line already flushed in the caller, treat buffer as a single paragraph.
-        Handle existing blank lines for natural paragraph separation.
-        """
-        if not buffer:
-            return []
-        paragraphs: List[str] = []
-        current: List[str] = []
-        for ln in buffer:
-            if ln.strip() == "":
-                if current:
-                    paragraphs.append("\n".join(current).strip())
-                    current = []
-            else:
-                current.append(ln)
-        if current:
-            paragraphs.append("\n".join(current).strip())
-        return paragraphs
-
-    def _parse_chapter_decl(self, line: str) -> Chapter:
-        """Format: ``@chapter id: Title`` or ``@chapter id``."""
-        content = line[len("@chapter"):].strip()
-        if ":" in content:
-            cid, title = content.split(":", 1)
-            return Chapter(chapter_id=cid.strip(), title=title.strip())
-        return Chapter(chapter_id=content.strip(), title=content.strip())
-
-    def _parse_branch_header(self, header_line: str, chapter_id: str) -> Branch:
-        """Format: ``# id: Title`` or ``# id``."""
-        content = header_line.lstrip("#").strip()
-        if ":" in content:
-            bid, title = content.split(":", 1)
-            return Branch(branch_id=bid.strip(), title=title.strip(), chapter_id=chapter_id)
-        else:
-            bid = content.strip()
-            return Branch(branch_id=bid, title=bid, chapter_id=chapter_id)
-
-    def _parse_choice_line(self, line: str) -> Choice:
-        """
-        Format: "* [condition] text -> target_id" or "* text -> target_id"
-        """
-        body = line[2:].strip()
-        if "->" not in body:
-            raise ParseError("Choice line must contain '->'.")
-        left, right = body.split("->", 1)
-        left = left.strip()
-        condition: Optional[str] = None
-        text: str
-        if left.startswith("["):
-            end = left.find("]")
-            if end == -1:
-                raise ParseError("Missing closing ']' in condition.")
-            condition = left[1:end].strip()
-            text = left[end + 1:].strip()
-        else:
-            text = left
-        target = right.strip()
-        if not text or not target:
-            raise ParseError("Choice text or target is empty.")
-        return Choice(text=text, target_id=target, condition=condition)
-
-    def _parse_action_line(self, line: str) -> Action:
-        content = line[1:].strip()
-        if content.startswith("set "):
-            rest = content[4:].strip()
-            if "=" not in rest:
-                raise ParseError("Invalid set syntax.")
-            var, val = rest.split("=", 1)
-            return Action(op="set", var=var.strip(), value=self._parse_value(val.strip()))
-        if content.startswith("add "):
-            rest = content[4:].strip()
-            if "+=" not in rest:
-                raise ParseError("Invalid add syntax.")
-            var, val = rest.split("+=", 1)
-            return Action(op="add", var=var.strip(), value=self._parse_value(val.strip()))
-        m = re.match(r"(\w+)\s*(=|\+=|-=|\*=|/=|//=|%=|\*\*=)\s*(.+)", content)
-        if m:
-            var, op, val = m.groups()
-            op_map = {
-                "=": "set",
-                "+=": "add",
-                "-=": "sub",
-                "*=": "mul",
-                "/=": "div",
-                "//=": "floordiv",
-                "%=": "mod",
-                "**=": "pow",
-            }
-            if op == "=":
-                try:
-                    parsed = self._parse_value(val.strip())
-                    return Action(op="set", var=var.strip(), value=parsed)
-                except ParseError:
-                    return Action(op="expr", var=var.strip(), value=val.strip())
-            return Action(op=op_map[op], var=var.strip(), value=self._parse_value(val.strip()))
-        raise ParseError("Unknown action command.")
-
-    def _parse_value(self, token: str) -> Union[int, float, bool, str]:
-        t = token.lower()
-        if t == "true":
-            return True
-        if t == "false":
-            return False
-        if (token.startswith('"') and token.endswith('"')) or (
-            token.startswith("'") and token.endswith("'")
-        ):
-            try:
-                return ast.literal_eval(token)
-            except Exception:
-                raise ParseError(f"Invalid value: {token}")
-        try:
-            return int(token)
-        except ValueError:
-            try:
-                return float(token)
-            except ValueError:
-                raise ParseError(f"Invalid value: {token}")
 
 
 @dataclass
